Anchor tracker-detail period history + pace on the local date, not UTC (#56)
`AnalyticsTrackersDetail` derived "today" from UTC in two places while the
rest of the trackers domain uses `localDateString()`:

- `buildCalendarPeriodHistory` passed a raw `new Date()` to
  `calendarPeriodBounds` (which reads a Date's UTC calendar fields), so a
  UTC+10/+11 user opening the page before ~10am local saw the previous
  calendar month/week as "Current" with every row's bounds shifted a day.
- the component-level `today` (`new Date().toISOString().slice(0,10)`) fed
  the pace math (`daysElapsed` / `daysLeft`) and the pace-chart `today`
  marker, one day stale in the same window.

Both now use the local calendar date; the `calendarPeriodBounds` call gets
it anchored at noon UTC (the idiom `dateStr.ts` uses internally).

`buildCalendarPeriodHistory` is extracted to a sibling
`trackerPeriodHistory.ts` so it can be unit-tested (matching
`importErrorClassification.ts` and the #49/#51 extraction). New test runs
under `TZ=Australia/Sydney` at `2026-03-31T14:00Z` (01:00 local, next local
day) and fails against the pre-fix UTC code.
